chore: update boilerplate configuration
