--- backend/.eslintrc.js
@@ -1,13 +1,0 @@
-module.exports = {
-  root: true,
-  env: {
-    jest: true,
-  },
-  extends: 'airbnb-base',
-  rules: {
-    'no-underscore-dangle': 0,
-    'no-param-reassign': 0,
-    'no-return-assign': 0,
-    camelcase: 0,
-  },
-};
