+++ backend/README.md
@@ -1,9 +1,9 @@
-# Express API Starter
+# Express API Starter with Typescript
 
 How to use this template:
 
 ```sh
-npx create-express-api --directory my-api-name
+npx create-express-api --typescript --directory my-api-name
 ```
 
 Includes API Server utilities:
@@ -19,10 +19,16 @@ Includes API Server utilities:
 
 Development utilities:
 
+* [typescript](https://www.npmjs.com/package/typescript)
+  * TypeScript is a language for application-scale JavaScript.
+* [ts-node](https://www.npmjs.com/package/ts-node)
+  * TypeScript execution and REPL for node.js, with source map and native ESM support.
 * [nodemon](https://www.npmjs.com/package/nodemon)
   * nodemon is a tool that helps develop node.js based applications by automatically restarting the node application when file changes in the directory are detected.
 * [eslint](https://www.npmjs.com/package/eslint)
   * ESLint is a tool for identifying and reporting on patterns found in ECMAScript/JavaScript code.
+* [typescript-eslint](https://typescript-eslint.io/)
+  * Tooling which enables ESLint to support TypeScript.
 * [jest](https://www.npmjs.com/package/jest)
   * Jest is a delightful JavaScript Testing Framework with a focus on simplicity.
 * [supertest](https://www.npmjs.com/package/supertest)
@@ -43,7 +49,7 @@ npm run lint
 ## Test
 
 ```
-npm test
+npm run test
 ```
 
 ## Development
