--- backend/src/api/emojis.js
@@ -1,9 +1,0 @@
-const express = require('express');
-
-const router = express.Router();
-
-router.get('/', (req, res) => {
-  res.json(['😀', '😳', '🙄']);
-});
-
-module.exports = router;
