+++ backend/test/api.test.ts
@@ -1,6 +1,6 @@
-const request = require('supertest');
+import request from 'supertest';
 
-const app = require('../src/app');
+import app from '../src/app';
 
 describe('GET /api/v1', () => {
   it('responds with a json message', (done) => {
@@ -9,7 +9,7 @@ describe('GET /api/v1', () => {
       .set('Accept', 'application/json')
       .expect('Content-Type', /json/)
       .expect(200, {
-        message: 'API - 👋🌎🌍🌏'
+        message: 'API - 👋🌎🌍🌏',
       }, done);
   });
 });
